+++ layers/Engine/packages/component-model/src/Resolvers/QueryableInterfaceSchemaDefinitionResolverAdapter.php
@@ -13,8 +13,10 @@ class QueryableInterfaceSchemaDefinitionResolverAdapter extends InterfaceSchemaD
 {
     public function getFieldFilterInputContainerComponent(ObjectTypeResolverInterface $objectTypeResolver, string $fieldName): ?Component
     {
-        /** @var QueryableInterfaceTypeFieldSchemaDefinitionResolverInterface */
-        $interfaceTypeFieldSchemaDefinitionResolver = $this->interfaceTypeFieldSchemaDefinitionResolver;
-        return $interfaceTypeFieldSchemaDefinitionResolver->getFieldFilterInputContainerComponent($fieldName);
+        if (!($this->interfaceTypeFieldSchemaDefinitionResolver instanceof QueryableInterfaceTypeFieldSchemaDefinitionResolverInterface)) {
+            return null;
+        }
+
+        return $this->interfaceTypeFieldSchemaDefinitionResolver->getFieldFilterInputContainerComponent($fieldName);
     }
 }